# SiteManager/public/swagger/swagger.yaml
openapi: 3.0.0
info:
  title: Site Manager API
  description: Site Manager API Documentation
  contact:
    email: ndungudennis250@gmail.com
  version: 1.0.0
servers:
  - url: http://172.105.90.112/site-manager-backend/SiteManager
    description: Production server
  - url: http://localhost
    description: Development server
paths:
...
  /api/clockIn:
    post:
      tags:
        - ClockIns
      summary: Clock in
      description: Clock in
      operationId: clockIn
      requestBody:
        description: Pass clock in details
        required: true
        content:
          application/json:
            schema:
              required:
                - siteManagerId
                - projectId
                - workerId
                - clockInTime
              properties:
                siteManagerId:
                  type: integer
                  format: int64
                  example: 1
                projectId:
                  type: integer
                  format: int64
                  example: 1
                workerId:
                  type: integer
                  format: int64
                  example: 1
                clockInTime:
                  type: string
                  format: date-time
                  example: '2023-07-20'
              type: object
      responses:
        '201':
          description: Clocked in successfully
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: Clocked in successfully
                type: object
        '404':
          description: Worker does not exist
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: Worker does not exist
                type: object
        '500':
          description: Server error
...